+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Added 7.2 resources
+## Added patch 7.2 resources
 
 - Added 996 new icons from patch 7.2 (for a total of 17,671 icons available.)
 - Added 19 new images from patch 7.2 to insert in your profiles.
@@ -12,3 +12,4 @@
 
 - Removed the quotes surrounding the "currently" text in the tooltip (IC and OOC). ([ticket #88](https://wow.curseforge.com/projects/total-rp-3/issues/88))
 - The profile sanitization option will now also crop unreasonably long texts. ([ticket #92](https://wow.curseforge.com/projects/total-rp-3/issues/92))
+- Removed code related to our April fools' day joke.